Fixed a bug in CreatedAccountDialog - isInputValid()
Method was looping over all users to check if all fields were supplied
and if passwords matched for the user that wasn't created yet. This also
had the bug that if you had no users, it wouldn't check any of those
conditions and you could created a user with all or some "blank" fields.