Changed SQL when creating tables and modified createTables() method
Adding "CREATE TABLE IF NOT EXISTS TABLENAME" removes the need to check
in the createTables() method if the Users table is empty before creating 
the other tables. This also enables one to add a DB that has the correct
tables, but is empty. Of a DB that is completely empty.